==> code/analysis/3-analysis.Rmd <==
---
title: "Descriptive analysis"
author: "Cole Tanigawa-Lau"
date: ''
output:
  html_document:
    theme: paper
    df_print: paged
    toc: yes
    toc_depth: 3
    toc_float: yes
editor_options:
  chunk_output_type: inline
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, fig.width = 7, fig.height = 5)
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
```

```{r about, echo=FALSE, results = "asis"}
cat(
  sprintf("Updated: %s.", 
          format(Sys.time(), "%b %d, %Y at %H:%M:%S", usetz = TRUE)),
  sprintf("Working directory: `%s`.", getwd()),
  sep = "\n\n"
)
```


Reads:

  - data/mod_RF_absolute.rds

Writes:

  - FILEPATH

# Setup

```{r packages, warning=FALSE, message=FALSE}
library(coler)
library(dplyr)
library(data.table)

library(caret)
library(randomForest)

library(stargazer)
```

```{r final prep, child = "0-final_prep.Rmd"}
source("code/analysis/0-final_prep.R")
```

```{r resampling}
# Up-sample
absolute_upsamp  <- absolute[ , upSample(.SD, closed)]
percap_upsamp    <- percap[   , upSample(.SD, closed)]
  
# Down-sample
absolute_downsamp  <- absolute[ , downSample(.SD, closed)]
percap_downsamp    <- percap[   , downSample(.SD, closed)]
```


```{r}
models <- readRDS("data/fitted_models.rds")
```


# Balance
```{r}
baltests <- absolute[ , map(.SD, ~ t.test(.x ~ closed, .SD) %>% list()), 
                      .SDcols = vars_absolute]

balDT <- 
  map(baltests, ~ .x[[1]] %>% broom::tidy()) %>% 
  rbindlist(idcol = "variable")

# estimate: estimate1 - estimate2 = treatment - control
setnames(balDT, c("estimate1", "estimate2"), c("mean_closed", "mean_open"))

balDT[ , 
       varname := c("White population", "Black population", "English proficiency",
                    "Aggregate income", "Public assistance income", "Social Security income", "Poverty",
                    "Seniors", "High school grads", "Population", 
                    "Computers", "No computer", "Internet", "Service area size", 
                    "Distance to nearest office", "Number of Census tracts") %>% 
         as.factor()
       ]
```


```{r}
setorder(balDT, -estimate)
balDT[ , varname := factor(varname, levels = varname)]
```

```{r}
plot_bal <- 
  ggplot(balDT) +
  geom_pointrange(aes(x = varname, y = estimate, 
                      ymin = conf.low, ymax = conf.high),
                  size = 0.2) +
  geom_hline(aes(yintercept = 0), linetype = "dashed") +
  labs(x = "Predictor", 
       y = latex2exp::TeX("$\\mu_{closed} - \\mu_{open}$")) +
  coord_flip() + theme_bw(10)

plot_bal
```

```{r}
ggsave(plot = plot_bal, filename = "writing/figures/diffmeans.pdf")
```


```{r results = "asis"}
round_num <- function(data, func = round, digits = 3){
  num_test <- vapply(data, is.numeric, logical(1))
  num_cols <- names(which(num_test))
  
  out <- dplyr::mutate_at(data, num_cols, func, digits = digits)
  
  return(out)
}

balDT[order(-abs(estimate))] %>% 
  round_num(func = signif, digits = 3) %>% 
  knitr::kable()
```


# K-Means
```{r}
kmeans_absolute <- kmeans(absolute[ , ..vars_absolute], centers = 2)
absolute[ , kluster := as.factor(kmeans_absolute$cluster - 1)]

absolute[ , confusionMatrix(kluster, reference = closed, positive = "1")]
```

# Logit
## Coefficients
### Fit model on standardized data
```{r}
absolute_std <- copy(absolute)
set_cols(absolute_std, j = vars_absolute, FUN = scale)

models$logit.absolute.std <- 
  glm(formula = form_absolute, family = binomial(link = "logit"), 
      data = absolute_std)
```

```{r baseline, results = "asis", eval = FALSE, include = FALSE}
logit_absolute <- 
  list(models$logit.absolute)

outcome_means <- map(logit_absolute, c("data", "closed")) %>% 
  map( ~ mean(as.numeric(as.character(.x)))) %>% unlist() %>% 
  round(3)


stargazer::stargazer(
  logit_absolute,
  
  # column.labels = c("Original", "Upsampled", "Downsampled"),
  # ci = TRUE, 
  # star.cutoffs = c(NULL, NA, NA), se = list(NULL, NA, NA),
  covariate.labels = c("White population", "Black population", "English proficiency",
                       "Aggregate income", "Social Security income", "Poverty",
                       "Seniors", "High school grads", "Population", 
                       "Computers", "No computer", "Internet", "Service area size", 
                       "Distance to nearest office", "Number of Census tracts"),
  
  add.lines = list(c("Prop. closed", outcome_means[1])),
  
  notes = "Standard errors in parentheses. Predictors were scaled by one standard deviation.",
  notes.append = FALSE, notes.align = "l",
  
  omit.stat = c("aic", "ll"), 
  type = "html", header = FALSE,
  out = "writing/tables/tab_logit.tex"
  )
```


```{r bias corrected, results = "asis", eval = FALSE, include = FALSE}
logit_absolute <- 
  list(models$logit.absolute,
       brglm::brglm(formula = form_absolute, family = binomial(link = "logit"),
               data = mutate(absolute_std, closed = as.numeric(as.character(closed)))))

brglm::brglm(formula = form_absolute, family = binomial(link = "logit"),
               data = mutate(absolute_std, closed = as.numeric(as.character(closed))))

outcome_means <- map(logit_absolute, c("data", "closed")) %>% 
  map( ~ mean(as.numeric(as.character(.x)))) %>% unlist() %>% 
  round(3)


stargazer::stargazer(
  logit_absolute,
  
  # column.labels = c("Original", "Upsampled", "Downsampled"),
  # ci = TRUE, 
  # star.cutoffs = c(NULL, NA, NA), se = list(NULL, NA, NA),
  covariate.labels = c("White population", "Black population", "English proficiency",
                       "Aggregate income", "Public assistance income", "Social Security income", "Poverty",
                       "Seniors", "High school grads", "Population", 
                       "Computers", "No computer", "Internet", "Service area size", 
                       "Distance to nearest office", "Number of Census tracts"),
  
  add.lines = list(c("Prop. closed", outcome_means[1])),
  
  notes = "Standard errors in parentheses. Predictors were scaled by one standard deviation.",
  notes.append = FALSE, notes.align = "l",
  
  omit.stat = c("aic", "ll"), 
  type = "html",
  out = "writing/tables/tab_logit.tex"
  )
```



```{r augment, results = "asis", eval = TRUE}
logit_absolute <- 
  list(models$logit.absolute,
       models$logit.absolute_upsamp,
       models$logit.absolute_downsamp)

outcome_means <- map(logit_absolute, c("data", "closed")) %>% 
  map( ~ mean(as.numeric(as.character(.x)))) %>% unlist() %>% 
  round(3)


stargazer::stargazer(
  logit_absolute,
  
  column.labels = c("Original", "Upsampled", "Downsampled"),
  # ci = TRUE, 
  star.cutoffs = c(NULL, NA, NA), se = list(NULL, NA, NA),
  covariate.labels = c(#"Constant", 
                       "White population", "Black population", "English proficiency",
                       "Aggregate income", "Public assistance income", "Social Security income", "Poverty",
                       "Seniors", "High school grads", "Population", 
                       "Computers", "No computer", "Internet", "Service area size", 
                       "Distance to nearest office", "Number of Census tracts"),
  
  add.lines = list(c("Prop. closed", outcome_means)),
  dep.var.labels = "Office closed",
  
  # notes = "Standard errors in parentheses. Predictors were scaled by one standard deviation. I omit uncertainty estimates for the up- and downsampled data sets because they are incomparable, and statistical significance there is meaningless.",
  notes = "",
  notes.append = FALSE,
  title = "Conditional expectations of office closure",
  
  omit.stat = c("aic", "ll"), 
  type = "html", font.size = "footnotesize",
  out = "writing/tables/tab_logit.tex"
  )
```


## Confusion matrix: fitted values {.tabset}

Full versions of the output in 2-eval script.

### Original sample

```{r}
confusionMatrix(
  
  models$logit.absolute$fitted.values %>% 
    cutoff(., threshold = median(.)),
  
  reference = absolute$closed,
  
  positive = "1"
)
```

### Upsample

```{r}
confusionMatrix(
  
  models$logit.absolute_upsamp$fitted.values %>% 
    cutoff(., threshold = median(.)),
  
  reference = absolute_upsamp$closed,
  
  positive = "1"
)
```

### Downsample

```{r}
confusionMatrix(
  
  models$logit.absolute_downsamp$fitted.values %>% 
    cutoff(., threshold = median(.)),
  
  reference = absolute_downsamp$closed,
  
  positive = "1"
)
```


# Random forest

## Variable importance {.tabset}

### Original sample
```{r}
importance_orig <- 
  models$RF.absolute$importance %>% 
  as_tibble(rownames = "variable")
 
arrange(importance_orig, -MeanDecreaseAccuracy)
```

### Upsampled
```{r}
importance_upsamp <- 
  models$RF.absolute_upsamp$importance %>% 
  as_tibble(rownames = "variable")


arrange(importance_upsamp, -MeanDecreaseAccuracy)
```

### Downsampled
```{r}
importance_downsamp <- 
  models$RF.absolute_downsamp$importance %>% 
  as_tibble(rownames = "variable")

arrange(importance_downsamp, -MeanDecreaseAccuracy)
```

### Comparison figure

```{r}
importances <- 
  rbindlist(list(Original = importance_orig, 
                 Upsample = importance_upsamp, 
                 Downsample = importance_downsamp), 
            idcol = "Sample")
```


```{r}
importances[ , Sample := as.factor(Sample)
             ][ , Sample := factor(Sample, levels = c("Original", "Upsample", "Downsample"))]


importances[balDT, varname := i.varname, on = "variable"]

importances[ , 
             varname := factor(varname, 
                               levels = .SD[Sample == "Original"
                                            ][order(MeanDecreaseAccuracy), unique(varname)])
             ]
```

```{r}
plot_importance <- 
  ggplot(importances) +
  geom_col(aes(x = varname, y = MeanDecreaseAccuracy),
           width = 0.25, position = "dodge") +
  facet_grid(. ~ Sample, scale = "free") +
  scale_y_continuous(n.breaks = 4) +
  labs(x = "Variable", y = "Relative importance (mean decrease in accuracy)") +
  theme_bw(10) + coord_flip()

plot_importance
```

```{r}
ggsave(plot = plot_importance, filename = "writing/figures/rf_importances.pdf")
```


## Confusion matrix {.tabset}

Full versions of the output in 2-eval script.


### Original sample

```{r}
confusionMatrix(predict(models$RF.absolute, newdata = absolute),
  
  reference = absolute$closed,
  
  positive = "1")
```

### Upsample

```{r}
confusionMatrix(predict(models$RF.absolute_upsamp, newdata = absolute),
  
  reference = absolute$closed,
  
  positive = "1")
```

```{r}
models$RF.absolute_upsamp
```


### Downsample

```{r}
confusionMatrix(predict(models$RF.absolute_downsamp, newdata = absolute),
  
  reference = absolute$closed,
  
  positive = "1")
```

```{r}
models$RF.absolute_downsamp
```


<!-- ### OOB predictions -->
<!-- ```{r} -->
<!-- # OOB prediction -->
<!-- confusionMatrix(models$RF.absolute$predicted, -->

<!--   reference = absolute$closed, -->

<!--   positive = "1") -->
<!-- ``` -->

# Out-of-sample prediction

```{r}
set.seed(575)
test <- sample_frac(absolute, 0.20)
train <- setdiff(absolute, test)
```

## Logit {.tabset}

### Original sample

```{r}
logit <- glm(data = train, 
             formula = form_absolute, 
             family = binomial(link = "logit"))

mean_out_train <- train[ , mean(as.numeric(as.character(closed)))]

test[ , caret::confusionMatrix(closed, 
                               predict(logit, newdata = test, type = "response") %>% 
                                 cutoff(threshold = mean_out_train),
                               
                               positive = "1")
      ]
```

### Upsample
```{r}
train_upsamp <- train[ , upSample(train, closed)] %>% as.data.table()

logit <- glm(data = train_upsamp, 
             formula = form_absolute, 
             family = binomial(link = "logit"))

mean_out_train <- train_upsamp[ , mean(as.numeric(as.character(closed)))]

test[ , caret::confusionMatrix(closed, 
                               predict(logit, newdata = test, type = "response") %>% 
                                 cutoff(threshold = mean_out_train),
                               
                               positive = "1")
      ]
```

### Downsample
```{r}
train_downsamp <- train[ , downSample(train, closed)] %>% as.data.table()

logit <- glm(data = train_downsamp, 
             formula = form_absolute, 
             family = binomial(link = "logit"))

mean_out_train <- train_downsamp[ , mean(as.numeric(as.character(closed)))]

test[ , caret::confusionMatrix(closed, 
                               predict(logit, newdata = test, type = "response") %>% 
                                 cutoff(threshold = mean_out_train),
                               
                               positive = "1")
      ]
```


## Random forest

Not really necessary since we have OOB performance.

```{r}
mtries <- 3:8
mods <- 
  parallel::mclapply(mtries,
                     FUN = function(mtry) 
                       randomForest(x = train[ , ..vars_absolute],
                                    y = train[ , closed], mtry = mtry),
                     mc.cores = length(mtries))

map(mods, ~ test[ , caret::confusionMatrix(closed, predict(.x, newdata = test),
                                           positive = "1")]) %>% 
  setNames(mtries)
```
